refactor: move changeTags procedure to admin router
