+++ src/server/api/routers/club.ts
@@ -21,12 +21,7 @@ import { userMetadataToClubs } from '@src/server/db/schema/users';
 import { syncCalendar, watchCalendar } from '@src/utils/calendar';
 import { createClubSchema } from '@src/utils/formSchemas';
 import { getGoogleAccessToken } from '@src/utils/googleAuth';
-import {
-  adminProcedure,
-  createTRPCRouter,
-  protectedProcedure,
-  publicProcedure,
-} from '../trpc';
+import { createTRPCRouter, protectedProcedure, publicProcedure } from '../trpc';
 import { clubEditRouter } from './clubEdit';
 
 const byNameSchema = z.object({
@@ -44,11 +39,6 @@ const bySlugSchema = z.object({
 
 const joinLeaveSchema = z.object({
   clubId: z.string().default(''),
-});
-
-const tagReplaceSchema = z.object({
-  oldTag: z.string(),
-  newTag: z.string(),
 });
 
 const searchSchema = z.object({
@@ -449,31 +439,6 @@ export const clubRouter = createTRPCRouter({
         console.error(e);
         throw e;
       }
-    }),
-  changeTags: adminProcedure
-    .input(tagReplaceSchema)
-    .mutation(async ({ input, ctx }) => {
-      const clubsToChange = await ctx.db.query.club.findMany({
-        where: sql`${input.oldTag} = ANY(tags)`,
-      });
-      clubsToChange.map((club) => {
-        club.tags = club.tags.map((tag) =>
-          tag == input.oldTag ? input.newTag : tag,
-        );
-        return club;
-      });
-      const clubPromise: Promise<unknown>[] = [];
-      for (const clu of clubsToChange) {
-        clubPromise.push(
-          ctx.db
-            .update(club)
-            .set({ tags: clu.tags })
-            .where(eq(club.id, clu.id)),
-        );
-      }
-      await Promise.all(clubPromise);
-      await ctx.db.refreshMaterializedView(usedTags);
-      return { affected: clubsToChange.length };
     }),
   tagSearch: publicProcedure
     .input(searchTagSchema)
